# Conflicts:
#	public/js/app.js
#	resources/js/main.js

+++ database/migrations/2020_09_14_026332_create_invoices_table.php
@@ -20,12 +20,12 @@ class CreateInvoicesTable extends Migration
             $table->date('start_date')->comment('Ngày khởi hành');
             $table->smallInteger('adult_count')->comment('Số lượng người lớn');
             $table->smallInteger('child_count')->comment('Số lượng trẻ em');
-            $table->float('sub_cost', 12, 3)->comment('Giá trị đơn hàng');
-            $table->float('vat_cost', 12, 3)->default(0)->comment('Thuế VAT');
-            $table->float('total_cost', 12, 3)->comment('Tổng giá trị');
+            $table->integer('sub_cost')->comment('Giá trị đơn hàng');
+            $table->integer('vat_cost')->default(0)->comment('Thuế VAT');
+            $table->integer('total_cost')->comment('Tổng giá trị');
             $table->string('payment_type')->comment('Hình thức thanh toán');
             $table->string('payment_code')->comment('Mã thanh toán app');
-            $table->float('deposit_cost', 12, 3)->comment('Chi phí đặt cọc (đã trả)');
+            $table->integer('deposit_cost')->comment('Chi phí đặt cọc (đã trả)');
 
             $table->string('customer_name')->comment('Tên khách hàng');
             $table->string('customer_address')->comment('Địa chỉ');
